rm pcl dep of laser_slam and moved declarations to laser_slam_ros
