+++ laser_slam/package.xml
@@ -19,5 +19,4 @@
   <depend>libpointmatcher</depend>
   <depend>mincurves</depend>
   <depend>minkindr</depend>
-  <depend>pcl_catkin</depend>
 </package>
